fix model and add post mapping

+++ src/main/java/com/sleeptracker/sleeptracker/repositories/SleepRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.sleeptracker.sleeptracker.models.Sleep;
 
 @Repository
-public interface SleepRepository extends CrudRepository<Sleep, Integer> {
+public interface SleepRepository extends CrudRepository<Sleep, Long> {
 
 }
